backend/requirements.txt
fastapi==0.78.0
motor==2.5.0
uvicorn==0.18.2
python-dotenv==0.19.2
pydantic==1.9.2
